vaultManager修正

+++ src/main/java/tutorial/test/hello_world/VaultManager.java
@@ -50,7 +50,7 @@ public class VaultManager {
     public void showBalance(UUID uuid){
         OfflinePlayer p = Bukkit.getOfflinePlayer(uuid).getPlayer();
         double money = getBalance(uuid);
-        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + (int)money);
+        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + money);
     }
     /////////////////////////////////////
     //      引き出し
@@ -64,7 +64,7 @@ public class VaultManager {
         EconomyResponse resp = economy.withdrawPlayer(p,money);
         if(resp.transactionSuccess()){
             if(p.isOnline()) {
-                p.getPlayer().sendMessage(ChatColor.YELLOW + "$" + (int) money + "支払いました");
+                p.getPlayer().sendMessage(ChatColor.YELLOW + "$" +  money + "支払いました");
             }
             return true;
         }
@@ -83,7 +83,7 @@ public class VaultManager {
         EconomyResponse resp = economy.depositPlayer(p,money);
         if(resp.transactionSuccess()){
             if(p.isOnline()){
-                p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+(int)money+"受取りました");
+                p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+money+"受取りました");
             }
             return true;
         }
